+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `citation-range-delimiter` term as an extension to CSL.
+
 ## [0.6.6] - 2024-11-18
 
 ### Added